+++ package.json
@@ -6,14 +6,13 @@
   "dependencies": {
     "@webrecorder/wombat": "^3.2.2",
     "brotli": "github:foliojs/brotli.js",
-    "fast-xml-parser": "^3.15.1",
+    "fast-xml-parser": "^3.19.0",
     "format-link-header": "^3.1.1",
     "http-status-codes": "^2.1.4",
-    "idb": "^5.0.1",
+    "idb": "^6.0.0",
     "js-levenshtein": "^1.1.6",
-    "js-yaml": "^3.13.1",
-    "json-2-csv": "^3.7.1",
-    "pako": "^1.0.10",
+    "js-yaml": "^4.1.0",
+    "pako": "^2.0.3",
     "parse-link-header": "^1.0.1",
     "parse5-html-rewriting-stream": "^6.0.1",
     "parse5-sax-parser": "^6.0.1",
